making the manager portal come true
